requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
aws-cdk-lib==2.87.0
aws-cdk.asset-awscli-v1==2.2.200
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v5==2.0.165
aws-cdk.aws-apigatewayv2-alpha==2.87.0a0
aws-cdk.aws-apigatewayv2-integrations-alpha==2.87.0a0
cattrs==23.1.2
certifi==2023.5.7
charset-normalizer==3.2.0
constructs==10.2.69
exceptiongroup==1.1.2
frozenlist==1.3.3
idna==3.4
importlib-resources==6.0.0
jsii==1.84.0
multidict==6.0.4
openai==0.27.8
publication==0.0.3
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
tqdm==4.65.0
typeguard==2.13.3
typing_extensions==4.7.1
urllib3==2.0.3
yarl==1.9.2
